Renamed button to kebab case
