+++ packages/wix-ui-core/src/components/Button-v2/button-v2.spec.tsx
@@ -1,10 +1,10 @@
 import * as React from "react";
 import { ReactDOMTestContainer } from "../../../test/dom-test-container";
-import { buttonDriverFactory } from "./buttonTwo.driver";
-import { buttonPrivateDriverFactory } from "./buttonTwo.driver.private";
-import ButtonTwo from "./buttonTwo";
+import { buttonDriverFactory } from "./button-v2.driver";
+import { buttonPrivateDriverFactory } from "./button-v2.driver.private";
+import { ButtonV2 } from "./";
 
-describe("ButtonTwo", () => {
+describe("ButtonV2", () => {
   const createDriver = new ReactDOMTestContainer()
     .unmountAfterEachTest()
     .createLegacyRenderer(buttonDriverFactory);
@@ -16,7 +16,7 @@ describe("ButtonTwo", () => {
   describe("onClick prop", () => {
     it("should be called on click", () => {
       const onClick = jest.fn();
-      const driver = createDriver(<ButtonTwo onClick={onClick} />);
+      const driver = createDriver(<ButtonV2 onClick={onClick} />);
       driver.click();
       expect(onClick).toBeCalled();
     });
@@ -26,13 +26,13 @@ describe("ButtonTwo", () => {
     const suffix = <div>suffix</div>;
     const prefix = <div>prefix</div>;
     it("should render suffix when given", () => {
-      const driver = createPrivateDriver(<ButtonTwo suffixIcon={suffix} />);
+      const driver = createPrivateDriver(<ButtonV2 suffixIcon={suffix} />);
       expect(driver.suffixExists()).toBeTruthy();
       expect(driver.prefixExists()).toBeFalsy();
     });
 
     it("should render prefix when given", () => {
-      const driver = createPrivateDriver(<ButtonTwo prefixIcon={prefix} />);
+      const driver = createPrivateDriver(<ButtonV2 prefixIcon={prefix} />);
       expect(driver.prefixExists()).toBeTruthy();
       expect(driver.suffixExists()).toBeFalsy();
     });
