+++ packages/wix-ui-core/src/components/Button-v2/index.ts
@@ -1,0 +1,1 @@
+export { ButtonV2 } from "./button-v2";
