--- packages/wix-ui-core/src/components/ButtonV2/index.ts
@@ -1,1 +1,0 @@
-export { ButtonTwo } from "./buttonTwo";
